Modified next-auth route config, as this had caused CI tests to fail.
