+++ frontend/src/app/api/auth/[...nextauth]/route.ts
@@ -1,27 +1,6 @@
 // app/api/auth/[...nextauth]/route.ts
-
 import NextAuth from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import { NextAuthOptions } from "next-auth";
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-  ],
-  callbacks: {
-    async session({ session, token }) {
-      // Add the user's Google ID to the session
-      if (session.user) {
-        session.user.id = token.sub as string;
-      }
-      return session;
-    },
-  },
-  secret: process.env.NEXTAUTH_SECRET,
-};
+import { authOptions } from "@/app/lib/authOptions";
 
 const handler = NextAuth(authOptions);
